chore: add license and code of conduct; expand codebook terms Add MIT LICENSE and Contributor Covenant Code of Conduct to the
repository to provide clear legal and community guidelines for users
and contributors.

Update codebook.toml to add several new terms ("behaviour",
"sexualized", "sublicense", "NONINFRINGEMENT") and fix punctuation in
the existing list. These additions align the codebook with the new
policy language introduced in the Code of Conduct and LICENSE files.